pyproject.toml
[tool.poetry]
name = "pi-mta-sign"
version = "0.1.0"
description = ""
authors = ["Lucas Oskorep <lucas.oskorep@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.31.0"
fastapi = "^0.100.0"
uvicorn = "^0.22.0"
python-dotenv = "^1.0.0"
fastapi-utils = "^0.2.1"
gtfs-realtime-bindings = "^1.0.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.0.277"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"




[tool.ruff]
# Enable the pycodestyle (`E`) and Pyflakes (`F`) rules by default.
# Unlike Flake8, Ruff doesn't enable pycodestyle warnings (`W`) or
# McCabe complexity (`C901`) by default.
select = ["E", "F"]
ignore = []

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]
# Same as Black.
line-length = 120

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Assume Python 3.10.
target-version = "py311"

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]